Forecast fetching and alert checking

+++ RainHound.Alerts/Models/FoundAlerts/FoundAlertModel.cs
@@ -4,7 +4,7 @@ public abstract class FoundAlertModel
     public string? Email { get; set; }
     public string? City { get; set; }
     public DateTime Hour { get; set; }
-    public decimal Threshold { get; set; }
-    public decimal Value { get; set; }
+    public double Threshold { get; set; }
+    public double Value { get; set; }
     public abstract string? Description { get; }
 }
